refactor(nx): include extensions in schema dts files
