+++ .eslintrc.yml
@@ -68,6 +68,7 @@ overrides:
   - files:
       - '*.d.mts'
       - '*.d.ts'
+      - '*.d.json.ts'
     rules:
       eslint-comments/disable-enable-pair: off
       eslint-comments/no-unlimited-disable: off
