+++ packages/@clc/nx/executors/esbuild/schema.d.json.ts
