+++ packages/@clc/nx/executors/package-json/schema.d.json.ts
