+++ packages/@clc/nx/executors/project-refs/schema.d.json.ts
