+++ packages/@clc/nx/executors/readme/schema.d.json.ts
